remove SignerConfigurator
